+++ CHANGELOG.md
@@ -41,6 +41,8 @@
 
 
 
+* Appsec
+    * Updated the `akamai_appsec_activations` resource to support moving hostnames between security configurations during activation.
 
 * Edgeworkers
   * Added possibility to fetch an `EdgeWorker` by name ([I#679](https://github.com/akamai/terraform-provider-akamai/issues/679)).